backend/requirements.txt
flask>=3.0
flask-cors>=4.0
gunicorn>=21.2
python-dotenv>=1.0
